move jsx tag auto close tests into jsx_tag_auto_close.rs
